corr test

+++ correct_test/rlibmOFA_float/allrep/LibTestHelper.h
@@ -134,8 +134,8 @@ unsigned long RunTestForExponent(int numExpBit) {
       double res = __ELEM__(x);
       
       for (int rnd_index = 0; rnd_index < 5; rnd_index++) {
-        double oracleResult = MpfrResult(x, numExpBit, bitlen, rnd_modes[rnd_index]);
-        double roundedRes = RoundDoubleToFEN(res, numExpBit, bitlen,
+        float oracleResult = MpfrResult(x, numExpBit, bitlen, rnd_modes[rnd_index]);
+        float roundedRes = RoundDoubleToFEN(res, numExpBit, bitlen,
                                              my_rnd_modes[rnd_index], 0);
 	
         if (oracleResult != oracleResult && roundedRes != roundedRes) continue;
